requirements.txt
numpy
pandas
python-dotenv
pyarrow
openai
pydantic
